make Bytes explicitly 64-bits

+++ cruncher/samples.go
@@ -6,7 +6,7 @@ type TCPFlowSample struct {
 	ID             string  `json:"id"`
 	Timestamp      string  `json:"timestamp"`
 	SampleDuration float64 `json:"sample-duration-s"`
-	Bytes          int     `json:"bytes"`
+	Bytes          int64   `json:"bytes"`
 	Bps            float64 `json:"bps"`
 	Retransmits    int     `json:"retransmits"`
 	SndCwnd        int     `json:"snd-cwnd"`
@@ -21,7 +21,7 @@ type UDPFlowSample struct {
 	ID             string  `json:"id"`
 	Timestamp      string  `json:"timestamp"`
 	SampleDuration float64 `json:"sample-duration-s"`
-	Bytes          int     `json:"bytes"`
+	Bytes          int64   `json:"bytes"`
 	Bps            float64 `json:"bps"`
 	JitterMs       float64 `json:"jitter-ms"`
 	LostPackets    int     `json:"lost-packets"`
